Move existing code into new package
