+++ src/main/scala/com/vlachjosef/shapeless/tagged-types-introduction.scala
@@ -1,4 +1,4 @@
-package com.vlachjosef
+package com.vlachjosef.shapeless
 
 package notypes {
   case class Arena(
